forgotPassword logic added
